Move extra tests in sub dirs
allows for proper typing
